Code refactoring v2
