+++ etna/datasets/__init__.py
@@ -4,6 +4,7 @@ from etna.datasets.datasets_generation import generate_from_patterns_df
 from etna.datasets.datasets_generation import generate_hierarchical_df
 from etna.datasets.datasets_generation import generate_periodic_df
 from etna.datasets.hierarchical_structure import HierarchicalStructure
+from etna.datasets.internal_datasets import load_dataset
 from etna.datasets.tsdataset import TSDataset
 from etna.datasets.utils import duplicate_data
 from etna.datasets.utils import set_columns_wide
